pyproject.toml
[project]
name = "fof-quant"
version = "0.1.0"
description = "ETF FOF research system for broad-index fund enhancement workflows."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "madeye" }]
dependencies = [
    "pydantic>=2.11",
    "pydantic-settings>=2.9",
    "pyyaml>=6.0",
    "typer>=0.15",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.15",
    "pytest>=8.3",
    "ruff>=0.11",
    "types-PyYAML>=6.0",
]

[project.scripts]
fof-quant = "fof_quant.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
